fix for year out of range; added unit test for that
